+++ src/pytest_ranking/plugin_utils.py
@@ -1,6 +1,6 @@
 # ------ Constants ------
 
-TCP_DATA_DIR = "tcp_data"
+DATA_DIR = "pytest_ranking_data"
 
 # Default amount of historical test run results to store per test
 DEFAULT_HIST_LEN = 50