+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add textual similarity between tests and changed files since last run as the third heuristic
 * Weight normalization
 * History length as an optional argument
+* Rename plugin to `pytest-ranking`
 
 
 0.1.0 (2023-12-04)